requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.20.0
Pillow>=8.4.0
easyocr>=1.5.0
streamlit>=1.10.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
numpy>=1.21.0
